+++ plugins/woocommerce-admin/client/task-list/test/index.js
@@ -92,6 +92,9 @@ describe( 'TaskDashboard and TaskList', () => {
 				type: 'setup',
 				action: 'CTA (optional)',
 				content: 'This is the optional task content',
+				additionalInfo: 'This is the task additional info',
+				expandable: true,
+				expanded: true,
 			},
 			{
 				key: 'required',
@@ -103,7 +106,9 @@ describe( 'TaskDashboard and TaskList', () => {
 				isDismissable: false,
 				type: 'setup',
 				action: 'CTA (required)',
+				actionLabel: 'This is the action label',
 				content: 'This is the required task content',
+				expandable: false,
 			},
 			{
 				key: 'completed',
@@ -730,25 +735,15 @@ describe( 'TaskDashboard and TaskList', () => {
 			},
 		] );
 		await act( async () => {
-			const { container, queryByText } = render(
-				<TaskDashboard query={ {} } />
-			);
+			const { queryByText } = render( <TaskDashboard query={ {} } /> );
 
 			// Expect the first incomplete task to be expanded
 			expect(
-				(
-					await findByText(
-						container,
-						'This is the optional task content'
-					)
-				 ).parentElement.style.maxHeight
-			).not.toBe( '0' );
+				queryByText( 'This is the optional task content' )
+			).not.toBeNull();
 
 			// Expect the second not to be.
-			expect(
-				queryByText( 'This is the required task content' ).parentElement
-					.style.maxHeight
-			).toBe( '0' );
+			expect( queryByText( 'This is the action label' ) ).toBeNull();
 		} );
 	} );