+++ plugins/woocommerce-admin/packages/experimental/src/experimental-list/collapsible-list/index.tsx
@@ -223,6 +223,10 @@ export const ExperimentalCollapsibleList: React.FC< CollapsibleListProps > = ( {
 		'woocommerce-experimental-list'
 	);
 
+	const wrapperClasses = classnames( {
+		'woocommerce-experimental-list-wrapper': ! isCollapsed,
+	} );
+
 	return (
 		<ExperimentalList { ...listProps } className={ listClasses }>
 			{ [
@@ -244,6 +248,7 @@ export const ExperimentalCollapsibleList: React.FC< CollapsibleListProps > = ( {
 						);
 						return (
 							<div
+								className={ wrapperClasses }
 								ref={ collapseContainerRef }
 								style={ transitionStyles }
 							>